requirements.txt
py-cord[speed]
python-dotenv